refactor: changed imports of styled components
